Introduce triton-arith-to-linalg pass (#85)
