+++ include/triton-shared/Conversion/TritonArithToLinalg/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(LLVM_TARGET_DEFINITIONS Passes.td)
+mlir_tablegen(Passes.h.inc -gen-pass-decls --name TritonArithToLinalg)
+add_public_tablegen_target(TritonArithToLinalgConversionPassIncGen)
