+++ lib/Conversion/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(TritonToLinalg)
 add_subdirectory(TritonToStructured)
+add_subdirectory(TritonArithToLinalg)
