+++ lib/Conversion/TritonArithToLinalg/CMakeLists.txt
@@ -1,0 +1,24 @@
+add_mlir_conversion_library(TritonArithToLinalg
+  TritonArithToLinalg.cpp
+  TritonArithToLinalgPass.cpp
+
+  DEPENDS
+  TritonArithToLinalgConversionPassIncGen
+
+  LINK_COMPONENTS
+  Core
+
+  LINK_LIBS PUBLIC
+  MLIRArithDialect
+  MLIRDialectUtils
+  MLIRIR
+  MLIRMathDialect
+  MLIRPass
+  MLIRTensorDialect
+  MLIRTransforms
+  MLIRSupport
+  TritonIR
+  TritonTransforms
+  TritonTilingExtIR
+  TritonStructuredIR
+)
